refactor(notes): update notes route import and remove obsolete test file
- Changed the import path for noteRoutes to use the new index.js structure.
- Deleted the outdated notes test file as it is no longer relevant to the current implementation.

Made-with: Cursor
